pyproject.toml
[tool.poetry]
name = "juno-api"
version = "0.1.0"
description = "API and DB for Juno"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.98.0"
uvicorn = "^0.22.0"
psycopg2 = "^2.9.6"
commit-linter = "^1.0.2"
pydantic = "^1.10.9"
sqlmodel = "^0.0.8"
SQLAlchemy = "1.4.41"
python-dotenv = "^1.0.0"
asyncpg = "^0.27.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.1"
pytest = "^7.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
